update des classes tests FieldMapping et Ajout de mappingRoad
