+++ force-app/main/default/lwc/mappingRoad/__tests__/mappingRoad.test.js
@@ -1,7 +1,7 @@
 import { createElement } from '@lwc/engine-dom';
-import MappingBoard from 'c/mappingBoard';
+import MappingRoad from 'c/mappingRoad';
 
-describe('c-mapping-board', () => {
+describe('c-mapping-road', () => {
     afterEach(() => {
         // The jsdom instance is shared across test cases in a single file so reset the DOM
         while (document.body.firstChild) {
@@ -11,8 +11,8 @@ describe('c-mapping-board', () => {
 
     it('TODO: test case generated by CLI command, please fill in test logic', () => {
         // Arrange
-        const element = createElement('c-mapping-board', {
-            is: MappingBoard
+        const element = createElement('c-mapping-road', {
+            is: MappingRoad
         });
 
         // Act
