+++ force-app/main/default/lwc/mappingRoad/mappingRoad.html
@@ -10,7 +10,7 @@
             Field Mapping (Drag & Drop)
           </h2>
         </div>
-      </header>
+      </header> 
     </div>
 
     <div class="slds-card__body slds-card__body_inner">
@@ -69,3 +69,4 @@
     </div>
   </article>
 </template>
+
