+++ force-app/main/default/lwc/mappingRoad/mappingRoad.js-meta.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-  <apiVersion>60.0</apiVersion>
-  <isExposed>true</isExposed>
+    <apiVersion>64.0</apiVersion>
+    <isExposed>true</isExposed>
   <targets>
     <target>lightning__AppPage</target>
     <target>lightning__RecordPage</target>
