+++ force-app/main/default/lwc/stepper/stepper.css
@@ -1,0 +1,54 @@
+.steps {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex: 1;
+  margin-left: 3rem;
+  position: relative;
+}
+.steps::before {
+  content: "";
+  position: absolute;
+  top: 20px;
+  left: 0;
+  right: 0;
+  height: 3px;
+  background: #ddd;
+  z-index: 0;
+}
+.steps div {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #f5f5f5;
+  border-radius: 50%;
+  width: 40px;
+  height: 40px;
+  margin-right: 100px;
+  font-weight: bold;
+  color: #999;
+  border: 2px solid #ddd;
+  position: relative;
+  z-index: 1;
+}
+.steps div::after {
+  content: attr(data-label);
+  font-size: 0.8rem;
+  margin-top: 0.5rem;
+  color: #666;
+  text-align: center;
+  font-weight: bolder;
+  white-space: nowrap;
+  position: absolute;
+  top: 50px;
+}
+.steps div.active {
+  background: #0070d2;
+  color: #fff;
+  border-color: #0070d2;
+}
+.steps div.completed {
+  background: #28a745;
+  color: #fff;
+  border-color: #28a745;
+}
