+++ force-app/main/default/lwc/stepper/stepper.html
@@ -1,0 +1,9 @@
+<template>
+  <div class="steps">
+    <div class={step1Class} data-label="Start">1</div>
+    <div class={step2Class} data-label="Select source">2</div>
+    <div class={step3Class} data-label="Mapping & transformation">3</div>
+    <div class={step4Class} data-label="Preview">4</div>
+    <div class={step5Class} data-label="Execution">5</div>
+  </div>
+</template>
